projects/obsidian-ticktick-sync/requirements.txt
# Core dependencies
requests>=2.31.0
watchdog>=3.0.0
pyyaml>=6.0.1
sqlalchemy>=2.0.23
python-dotenv>=1.0.0
click>=8.1.7
rich>=13.7.0

# Date/time handling
python-dateutil>=2.8.2

# Type hints
typing-extensions>=4.8.0

# Development dependencies
pytest>=7.4.3
pytest-cov>=4.1.0
pytest-mock>=3.12.0
black>=23.11.0
flake8>=6.1.0
mypy>=1.7.1

# Optional enhancements
aiohttp>=3.9.1  # For async operations (future)
